Refs #5 -- fix everything except with-redef edge case
The edge case is thus:

(trace #'blah)
(with-redefs [blah ...]
  (untrace #'blah))

When the with-redefs exits, it will restore the var root binding of
 #'blah to its state at the time the with-redefs was entered--which
means it will still be wrapped in the trace machinery. Worse, #'blah
will have been removed from the list of traced-vars, so subsequent
untraces won't work, and subsequent traces will leave #'blah with two
levels of trace reporting!